measurement_navigator/pyproject.toml
[tool.poetry]
name = "measurement_navigator"
version = "0.1.0"
description = ""
authors = ["Gibsy <gibsyvngr@gmail.com>"]

[tool.poetry.dependencies]
python = "~3.8"
amqp = "^5.1.1"
argon2-cffi = "^21.3.0"
argon2-cffi-bindings = "^21.2.0"
asgiref = "^3.5.2"
asttokens = "^2.0.7"
async-timeout = "^4.0.2"
attrs = "^22.1.0"
autopep8 = "^1.7.0"
backcall = "^0.2.0"
"backports.zoneinfo" = "^0.2.1"
beautifulsoup4 = "^4.11.1"
bleach = "^5.0.1"
boto3 = "^1.24.48"
botocore = "^1.27.48"
cachetools = "^5.2.0"
cchardet = "^2.1.7"
celery = "^5.2.7"
certifi = "^2022.6.15"
cffi = "^1.15.1"
chardet = "^5.0.0"
charset-normalizer = "^2.1.0"
click = "^8.1.3"
click-didyoumean = "^0.3.0"
click-plugins = "^1.1.1"
click-repl = "^0.2.0"
colorclass = "^2.2.2"
coreapi = "^2.3.3"
coreschema = "^0.0.4"
cryptography = "^37.0.4"
cycler = "^0.11.0"
debugpy = "^1.6.2"
decorator = "^5.1.1"
defusedxml = "^0.7.1"
Django = "4.1"
django-ckeditor = "^6.4.2"
django-cors-headers = "^3.13.0"
django-datatables-view = "^1.20.0"
django-debug-toolbar = "^3.5.0"
django-environ = "^0.9.0"
django-filter = "^22.1"
django-js-asset = "^2.0.0"
django-model-utils = "^4.2.0"
django-sass = "^1.1.0"
django-timezone-field = "^5.0"
django-tinymce = "^3.4.0"
django-webpack-loader = "^1.6.0"
django-widget-tweaks = "^1.4.12"
django-wysiwyg = "^0.8.0"
djangorestframework = "^3.13.1"
docopt = "^0.6.2"
drf-yasg = "^1.21.3"
entrypoints = "^0.4"
executing = "^0.9.1"
fastjsonschema = "^2.16.1"
flower = "^1.2.0"
fonttools = "^4.34.4"
frozenlist = "^1.3.1"
geoip2 = "^4.6.0"
aiohttp = "^3.8.1"
aiosignal = "^1.2.0"
django-celery-beat = "^2.3.0"
django-celery-results = "^2.4.0"
google-api-core = "^2.8.2"
google-auth = "^2.10.0"
google-cloud-bigquery = "^3.3.1"
google-cloud-bigquery-storage = "^2.14.1"
google-cloud-core = "^2.3.2"
google-crc32c = "^1.3.0"
google-resumable-media = "^2.3.3"
googleapis-common-protos = "^1.56.4"
grpcio = "^1.48.0"
grpcio-status = "^1.48.0"
gunicorn = "^20.1.0"
humanize = "^4.3.0"
idna = "^3.3"
importlib-metadata = "^4.12.0"
importlib-resources = "^5.9.0"
inflection = "^0.5.1"
ipykernel = "^6.15.1"
ipython = "^8.4.0"
ipython_genutils = "^0.2.0"
ipywidgets = "^7.7.1"
iso3166 = "^2.1.1"
itypes = "^1.2.0"
jedi = "^0.18.1"
Jinja2 = "^3.1.2"
jmespath = "^1.0.1"
jsonschema = "^4.9.1"
jupyter = "^1.0.0"
jupyter-client = "^7.3.4"
jupyter-console = "^6.4.4"
jupyter-core = "^4.11.1"
jupyterlab-pygments = "^0.2.2"
jupyterlab-widgets = "^1.1.1"
kiwisolver = "^1.4.4"
kombu = "^5.2.4"
libsass = "^0.21.0"
lz4 = "^4.0.2"
Markdown = "^3.4.1"
MarkupSafe = "^2.1.1"
matplotlib = "^3.5.2"
matplotlib-inline = "^0.1.3"
maxminddb = "^2.2.0"
multidict = "^6.0.2"
nbclient = "^0.6.6"
nbconvert = "^6.5.2"
nbformat = "^5.4.0"
nest-asyncio = "^1.5.5"
networkx = "^2.8.5"
notebook = "^6.4.12"
numpy = "^1.23.1"
oauthlib = "^3.2.0"
packaging = "^21.3"
pandas = "^1.4.3"
pandocfilters = "^1.5.0"
parso = "^0.8.3"
patsy = "^0.5.2"
pexpect = "^4.8.0"
pickleshare = "^0.7.5"
Pillow = "^9.2.0"
pip-upgrader = "^1.4.15"
prometheus-client = "^0.14.1"
prompt-toolkit = "^3.0.30"
proto-plus = "^1.22.0"
psutil = "^5.9.1"
psycopg2-binary = "^2.9.3"
ptyprocess = "^0.7.0"
pure-eval = "^0.2.2"
pyarrow = "^9.0.0"
pyasn1 = "^0.4.8"
pyasn1-modules = "^0.2.8"
pycodestyle = "^2.9.1"
pycparser = "^2.21"
pydot = "^1.4.2"
Pygments = "^2.12.0"
PyJWT = "^2.4.0"
pyparsing = "^3.0.9"
pyrsistent = "^0.18.1"
python-crontab = "^2.6.0"
python-dateutil = "^2.8.2"
python3-openid = "^3.2.0"
pytz = "^2022.1"
PyYAML = "^6.0"
pyzmq = "^23.2.0"
qtconsole = "^5.3.1"
QtPy = "^2.2.0"
requests = "^2.28.1"
requests-oauthlib = "^1.3.1"
rsa = "^4.9"
"ruamel.yaml" = "^0.17.21"
"ruamel.yaml.clib" = "^0.2.6"
s3transfer = "^0.6.0"
scipy = "^1.9.0"
Send2Trash = "^1.8.0"
six = "^1.16.0"
soupsieve = "^2.3.2"
sqlparse = "^0.4.2"
stack-data = "^0.3.0"
statsmodels = "^0.13.2"
swagger-spec-validator = "^2.7.4"
terminado = "^0.15.0"
terminaltables = "^3.1.10"
testpath = "^0.6.0"
tinycss2 = "^1.1.1"
toml = "^0.10.2"
tornado = "^6.2"
traitlets = "^5.3.0"
typing-extensions = "^4.3.0"
ujson = "^5.4.0"
uritemplate = "^4.1.1"
urllib3 = "^1.26.11"
vine = "^5.0.0"
wcwidth = "^0.2.5"
webencodings = "^0.5.1"
widgetsnbextension = "^3.6.1"
yarl = "^1.8.1"
zipp = "^3.8.1"
viztracer = "^0.15.3"
django-extensions = "^3.2.0"
django-timescaledb = "^0.2.12"
redis = "^4.3.4"
django-prometheus = "2.2.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

